+++ controllers/storage/scale.go
@@ -1,0 +1,7 @@
+package storage
+
+import "github.com/goharbor/harbor-cluster-operator/lcm"
+
+func (m *MinIOReconciler) Scale() (*lcm.CRStatus, error) {
+	panic("implement me")
+}
