+++ controllers/storage/update.go
@@ -1,0 +1,9 @@
+package storage
+
+import (
+	"github.com/goharbor/harbor-cluster-operator/lcm"
+)
+
+func (m *MinIOReconciler) Update() (*lcm.CRStatus, error) {
+	panic("implement me")
+}
